refactor: 다른 Screen 도 분리한 컴포넌트 사용
