+++ app/src/androidTest/java/nextstep/shoppingcart/CartScreenTest.kt
@@ -7,7 +7,8 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithContentDescription
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
-import nextstep.shoppingcart.ui.ShoppingCartScreen
+import nextstep.shoppingcart.data.Cart
+import nextstep.shoppingcart.ui.cart.ShoppingCartScreen
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
