+++ app/src/main/java/nextstep/shoppingcart/ProductDetailActivity.kt
@@ -9,7 +9,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import nextstep.shoppingcart.ui.ProductDetailScreen
+import nextstep.shoppingcart.ui.detail.ProductDetailScreen
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 class ProductDetailActivity : ComponentActivity() {
