+++ app/src/main/java/nextstep/shoppingcart/ShoppingCartActivity.kt
@@ -8,7 +8,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import nextstep.shoppingcart.ui.ShoppingCartScreen
+import nextstep.shoppingcart.ui.cart.ShoppingCartScreen
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 class ShoppingCartActivity : ComponentActivity() {
