+++ app/src/main/java/nextstep/shoppingcart/ui/productlist/ProductInfo.kt
@@ -1,22 +1,20 @@
-package nextstep.shoppingcart.ui.component
+package nextstep.shoppingcart.ui.productlist
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil.compose.AsyncImage
 import nextstep.shoppingcart.Product
-import nextstep.shoppingcart.R
+import nextstep.shoppingcart.ui.component.PriceLabel
+import nextstep.shoppingcart.ui.component.ProductImage
+import nextstep.shoppingcart.ui.component.ProductTitle
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 @Composable
@@ -31,22 +29,20 @@ fun ProductInfo(
         },
         horizontalAlignment = Alignment.Start,
     ) {
-        AsyncImage(
-            model = product.imageUrl,
+        ProductImage(
+            imageUrl = product.imageUrl,
             contentDescription = product.name,
-            contentScale = ContentScale.Crop,
             modifier = Modifier
                 .fillMaxWidth()
-                .height(158.dp)
+                .height(158.dp),
+            contentScale = ContentScale.Crop
         )
-        Text(
-            text = product.name,
-            style = MaterialTheme.typography.titleMedium,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis
+        ProductTitle(
+            title = product.name,
+            style = MaterialTheme.typography.titleMedium
         )
-        Text(
-            text = stringResource(id = R.string.price_format, product.price),
+        PriceLabel(
+            price = product.price,
             style = MaterialTheme.typography.labelLarge
         )
     }
