+++ app/src/main/java/nextstep/shoppingcart/ui/productlist/ProductListScreen.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.ui
+package nextstep.shoppingcart.ui.productlist
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.fillMaxSize
@@ -13,7 +13,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import nextstep.shoppingcart.Product
 import nextstep.shoppingcart.R
-import nextstep.shoppingcart.ui.component.ProductInfo
 import nextstep.shoppingcart.ui.component.ShoppingCartActionsTopBar
 
 @Composable
